+++ .simpspector.yml
@@ -22,4 +22,6 @@ doc_block:
   missing_newline_before_return: notice
   obsolete_variable: notice
   missing_variable: notice
+  missing_type_in_docblock: notice
+  missing_type_in_signature: notice
   type_missmatch: notice
